rename colors.ts to mods-colors.ts
