+++ src/components/filterPopovers/listSelect.tsx
@@ -1,6 +1,6 @@
 import type { Dispatch, SetStateAction } from "react";
 import { MultiSelect, createStyles } from "@mantine/core";
-import { colorsForDifficultyIndex } from "~/styles/colors";
+import { colorsForDifficultyIndex } from "~/styles/mods-colors";
